requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
langchain==0.0.350
gradio==3.50.0